refactor(realtime): rename sse bus and align docs with socket.io claims
